Added VS2005 project files and changes necessary for compiling in VS2005
git-svn-id: http://svn.nate.deepcreek.org.au/svn/ScanChannelsBDA/trunk@34 4d34ff29-007d-f442-978e-119f7a137983

+++ src/BDACard.cpp
@@ -294,7 +294,7 @@ HRESULT BDACard::GetSignalStatistics(BOOL &locked, BOOL &present, long &strength
 		return (log << "Cannot get node types\n").Show(hr);
 	}
 
-	for ( int i=0 ; i<NodeTypes ; i++ )
+	for ( int i=0 ; i<(int)NodeTypes ; i++ )
 	{
 		hr = bdaNetTop->GetControlNode(0, 1, NodeType[i], &iNode);
 		if (hr == S_OK)
